Success in validate url
Use url module, by requiring in it.,
By this module we can get if urlString is url then we will get all its object readily by parsing it by parse(urlString).objectName.
So one of that object , ie, hostname is we get.
That host name is the first argument for the dns.lookup() method and second argument is callback function which will give either err or address of that url hostname.

Note: -
Before using dns lookup method use reg expression to check if urlString  has http:// or https:// written or not at starting of urlString because we require url in format like https://www.example.com