Connect app to mongoDB
We connect this app to mongoDB, so that we can save our urls and shorturl in database
For that we require mongoose package.
And save url in .env file (by creating new .env file) and in that file save the mondoDB url which will connet to database and remember no space when savin that url between url string and "=" sign and that variable name is also in capital letters with no space.
Then connect that mongoDb variable to app by mongoose.connect() method.